fix: replace generic interface with refactored component
